Update to Qtile and git repository docs.
My Qtile now has an autostart.sh file it executes only once on startup
via a python hook. Additionally, many changes were added to the repo's
documentation. The program-install-list file is now in markdown and has
a more orderly list of software. There is also a very rudimentary script
for deploying the dotfiles to .config on a system. Its just a one line
cp command.

- Installed programs are now tracked via a markdown file
	- Deleted program-install-list.txt
	- Created program-install-list.md to track installed items

- Added ability for Qtile to reload wallpaper
	- Updated Qtile dotfiles to call autostart.sh script

- Created autostart.sh script under Qtile configs

- Created deploy-config.sh which deploys configs via copying.